conform to norm (mostly)
