+++ src/parser/parser_utils.c
@@ -1,5 +1,13 @@
 #include "shell.h"
 
+void	default_in_out(int *in, int *out, struct s_state *state)
+{
+	if (!state->in_set)
+		*in = 0;
+	if (!state->out_set)
+		*out = 1;
+}
+
 void	reset_cmd(char **command, char ***options, int *in, int *out)
 {
 	command[0] = NULL;
@@ -21,23 +29,6 @@ t_command	*create_cmd(char *command, char **options, int in, int out)
 	cmd->in = in;
 	cmd->out = out;
 	return (cmd);
-}
-
-struct s_state	init_state(void)
-{
-	struct s_state	state;
-
-	state.pipe = false;
-	state.option = false;
-	state.command = false;
-	state.redir_in = false;
-	state.redir_out = false;
-	state.append = false;
-	state.last = false;
-	state.here_doc = false;
-	state.in_set = false;
-	state.out_set = false;
-	return (state);
 }
 
 void	free_cmd(void *pointer)
